Move fetchWorkspaceTable into useAdminOrder hook
