+++ src/hooks/admin/useAdminOrder.tsx
@@ -1,11 +1,14 @@
 import useApi from '@hooks/useApi';
-import { Order, OrderStatus } from '@@types/index';
+import { Order, OrderStatus, PaginationResponse, Table } from '@@types/index';
 import { useSetRecoilState } from 'recoil';
-import { ordersAtom } from '@recoils/atoms';
+import { ordersAtom, tablePaginationResponseAtom } from '@recoils/atoms';
+import { useSearchParams } from 'react-router-dom';
 
 function useAdminOrder(workspaceId: string | undefined) {
   const { adminApi } = useApi();
   const setOrders = useSetRecoilState(ordersAtom);
+  const setTablePaginationResponse = useSetRecoilState(tablePaginationResponseAtom);
+  const [searchParams, setSearchParams] = useSearchParams();
 
   const fetchAllOrders = () => {
     adminApi.get<Order[]>('/orders', { params: { workspaceId } }).then((response) => {
@@ -58,7 +61,27 @@ function useAdminOrder(workspaceId: string | undefined) {
     });
   };
 
-  return { fetchAllOrders, payOrder, serveOrder, cancelOrder, fetchTodayOrders: fetchRealTimeOrders, fetchOrders, updateOrderProductServe, refundOrder };
+  const fetchWorkspaceTable = (tableNumber: number, page: number, size: number) => {
+    const params = { workspaceId, tableNumber, page, size };
+
+    adminApi.get<PaginationResponse<Table>>('/orders/table', { params }).then((res) => {
+      setTablePaginationResponse(res.data);
+      searchParams.set('page', params.page.toString());
+      setSearchParams(searchParams);
+    });
+  };
+
+  return {
+    fetchAllOrders,
+    payOrder,
+    serveOrder,
+    cancelOrder,
+    fetchTodayOrders: fetchRealTimeOrders,
+    fetchOrders,
+    updateOrderProductServe,
+    refundOrder,
+    fetchWorkspaceTable,
+  };
 }
 
 export default useAdminOrder;
