+++ src/hooks/admin/useAdminWorkspace.tsx
@@ -1,14 +1,13 @@
 import useApi from '@hooks/useApi';
 import { useSetRecoilState } from 'recoil';
-import { adminWorkspaceAtom, tablePaginationResponseAtom } from '@recoils/atoms';
-import { PaginationResponse, Table, Workspace } from '@@types/index';
-import { useNavigate, useSearchParams } from 'react-router-dom';
+import { adminWorkspaceAtom } from '@recoils/atoms';
+import { Workspace } from '@@types/index';
+import { useNavigate } from 'react-router-dom';
 
 function useAdminWorkspace() {
   const { adminApi } = useApi();
-  const [searchParams, setSearchParams] = useSearchParams();
   const setAdminWorkspace = useSetRecoilState(adminWorkspaceAtom);
-  const setTablePaginationResponse = useSetRecoilState(tablePaginationResponseAtom);
+
   const navigate = useNavigate();
 
   const fetchWorkspace = (workspaceId: string | undefined | null) => {
@@ -36,17 +35,7 @@ function useAdminWorkspace() {
       });
   };
 
-  const fetchWorkspaceTable = (workspaceId: number, tableNumber: number, page: number, size: number) => {
-    const params = { workspaceId, tableNumber, page, size };
-
-    adminApi.get<PaginationResponse<Table>>('/orders/table', { params }).then((res) => {
-      setTablePaginationResponse(res.data);
-      searchParams.set('page', params.page.toString());
-      setSearchParams(searchParams);
-    });
-  };
-
-  return { fetchWorkspace, updateWorkspaceTableCount, fetchWorkspaceTable };
+  return { fetchWorkspace, updateWorkspaceTableCount };
 }
 
 export default useAdminWorkspace;
